비지니스 로직 분리
